Added support for patching scandir module
- module contains backported os.scandir
